refactor(core): share URL validation
